+++ NEWS.md
@@ -1,3 +1,8 @@
+# roboplotr 2.8.2
+
+* Added params `set_external_menu(selected, max_length)` to control the initial selection and maximum length of items shown in the external menu created with `set_external_menu()`.
+* Fixes a `robolot(updatemenu)` issue on render.
+
 # roboplotr 2.8.1.1
 
 * `set_external_menu()` documentation added